Refactor AdminNoteListItem
